Fix slide matching for png, gif, and bmp files.
I have QA trivia slides that were not finding their way into the content database despite following the default format. It turns out that both the question and answer slides were matching the default question regular expression since they included the string 'png'. The | part of the regular expressions needs to be inside a group containing only the different extensions in order to work as desired.